+++ GameEngine/Collider.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "Math.h"
+#include "ObjectTransform.h"
 #include "Mesh.h"
 
 #define ICE_KINETIC_FRICTION 0.03
