+++ GameEngine/GameEngine.cpp
@@ -19,10 +19,7 @@ FILE* new_stdout;
 auto __ = freopen_s(&new_stdout, "CONOUT$", "w", stdout);
 #endif
 
-static float RotSpeedX = 0.0f;
-static float RotSpeedY = 30.0f;
-static float RotSpeedZ = 0.0f;
-static bool ShowStrs = false;
+static bool ShowStrs = true;
 static float FTheta = 0;
 static int CurrMesh = 0;
 static bool IsFullScreen = true;
@@ -35,19 +32,19 @@ std::string CWD = cwd;
 
 //DoTick_T Draw(GdiPP& Gdi, WndCreatorW& Wnd, const float& ElapsedTime, std::vector<Renderable>* ToRender, std::vector<LightObject*>* Lights)
 //{
-//	static float PrevFov = Cam.Fov;
+//	static float PrevFov = Cam->Fov;
 //	
 //	if (!LockCamera)
 //	{
-//		Cam.ViewAngles += Vec3((float)TerraPGE::DeltaMouse.y, 0, 0);
-//		Cam.ViewAngles.x = std::clamp<float>(Cam.ViewAngles.x, -89.0f, 89.0f);
-//		Cam.ViewAngles -= Vec3(0, (float)TerraPGE::DeltaMouse.x, 0);
+//		Cam->ViewAngles += Vec3((float)TerraPGE::DeltaMouse.y, 0, 0);
+//		Cam->ViewAngles.x = std::clamp<float>(Cam->ViewAngles.x, -89.0f, 89.0f);
+//		Cam->ViewAngles -= Vec3(0, (float)TerraPGE::DeltaMouse.x, 0);
 //	}
 //
-//	if (PrevFov != Cam.Fov)
+//	if (PrevFov != Cam->Fov)
 //	{
-//		Cam.CalcProjectionMat();
-//		PrevFov = Cam.Fov;
+//		Cam->CalcProjectionMat();
+//		PrevFov = Cam->Fov;
 //	}
 //
 //	if (GetAsyncKeyState(VK_INSERT) & 0x8000)
@@ -65,7 +62,7 @@ std::string CWD = cwd;
 //			IsFullScreen = !IsFullScreen;
 //		}
 //		TerraPGE::UpdateScreenInfo(Gdi);
-//		Cam.AspectRatio = (float)((float)TerraPGE::sy / (float)TerraPGE::sx);
+//		Cam->AspectRatio = (float)((float)TerraPGE::sy / (float)TerraPGE::sx);
 //	}
 //
 //	if (GetAsyncKeyState(VK_DELETE) & 0x8000)
@@ -80,47 +77,47 @@ std::string CWD = cwd;
 //		LockCamera = !LockCamera;
 //		Matrix Proj;
 //		Proj.MakeOrthoMatrix(-40.0f, 40.0f, -40.0f, 40.0f, TerraPGE::FNEAR, TerraPGE::FFAR);
-//		Cam.ProjectionMatrix = Proj;
-//		Cam.ViewMatrix = Matrix::CalcViewMatrix(((Vec3(0.0f, 0.0f, 0.0f) - LightSrcPos).Normalized()) * 100.0f, Vec3(0.0f, 0.0f, 0.0f), Vec3(0, 1, 0));
-//		Cam.Pos = -((Vec3(0.0f, 0.0f, 0.0f) - LightSrcPos).Normalized()) * 100.0f;
+//		Cam->ProjectionMatrix = Proj;
+//		Cam->ViewMatrix = Matrix::CalcViewMatrix(((Vec3(0.0f, 0.0f, 0.0f) - LightSrcPos).Normalized()) * 100.0f, Vec3(0.0f, 0.0f, 0.0f), Vec3(0, 1, 0));
+//		Cam->Pos = -((Vec3(0.0f, 0.0f, 0.0f) - LightSrcPos).Normalized()) * 100.0f;
 //	}
 //
 //	if (GetAsyncKeyState(VK_SPACE) & 0x8000)
 //	{
-//		Cam.Pos += (Cam.GetNewVelocity(Vec3(0, 1, 0) * Cam.CamRotation)) * ElapsedTime;
+//		Cam->Pos += (Cam->GetNewVelocity(Vec3(0, 1, 0) * Cam->CamRotation)) * ElapsedTime;
 //	}
 //
 //	if (GetAsyncKeyState(VK_LSHIFT) & 0x8000)
 //	{
-//		Cam.Pos += (Cam.GetNewVelocity(Vec3(0, -1, 0) * Cam.CamRotation)) * ElapsedTime;
+//		Cam->Pos += (Cam->GetNewVelocity(Vec3(0, -1, 0) * Cam->CamRotation)) * ElapsedTime;
 //	}
 //
 //	if (GetAsyncKeyState(VK_A) & 0x8000)
 //	{
-//		Cam.Pos += (Cam.GetNewVelocity(Vec3(-1, 0, 0) * Cam.CamRotation)) * ElapsedTime;
+//		Cam->Pos += (Cam->GetNewVelocity(Vec3(-1, 0, 0) * Cam->CamRotation)) * ElapsedTime;
 //	}
 //
 //	if (GetAsyncKeyState(VK_D) & 0x8000)
 //	{
-//		Cam.Pos += (Cam.GetNewVelocity(Vec3(1, 0, 0) * Cam.CamRotation)) * ElapsedTime;
+//		Cam->Pos += (Cam->GetNewVelocity(Vec3(1, 0, 0) * Cam->CamRotation)) * ElapsedTime;
 //	}
 //
 //	if (GetAsyncKeyState(VK_W) & 0x8000)
 //	{
-//		Cam.Pos += (Cam.GetNewVelocity(Vec3(0, 0, 1) * Cam.CamRotation)) * ElapsedTime;
+//		Cam->Pos += (Cam->GetNewVelocity(Vec3(0, 0, 1) * Cam->CamRotation)) * ElapsedTime;
 //	}
 //
 //	if (GetAsyncKeyState(VK_S) & 0x8000)
 //	{
-//		Cam.Pos += (Cam.GetNewVelocity(Vec3(0, 0, -1) * Cam.CamRotation)) * ElapsedTime;
+//		Cam->Pos += (Cam->GetNewVelocity(Vec3(0, 0, -1) * Cam->CamRotation)) * ElapsedTime;
 //	}
 //
 //	if (!LockCamera)
 //	{
-//		Cam.CamRotation = Matrix::CreateRotationMatrix(Cam.ViewAngles); // Pitch Yaw Roll
-//		Cam.LookDir = Cam.InitialLook * Cam.CamRotation;
-//		Vec3 T = Cam.Pos + Cam.LookDir;
-//		Cam.CalcCamViewMatrix(T);
+//		Cam->CamRotation = Matrix::CreateRotationMatrix(Cam->ViewAngles); // Pitch Yaw Roll
+//		Cam->LookDir = Cam->InitialLook * Cam->CamRotation;
+//		Vec3 T = Cam->Pos + Cam->LookDir;
+//		Cam->CalcCamViewMatrix(T);
 //	}
 //
 //	FTheta += 1.0f * ElapsedTime;
@@ -193,8 +190,8 @@ std::string CWD = cwd;
 //		Gdi.DrawStringA(20, 180, TriLinesStr + std::to_string(TerraPGE::ShowTriLines), RGB(255, 0, 0), TRANSPARENT);
 //		//Gdi.DrawStringA(20, 200, MeshStr + Meshes.at(CurrMesh).MeshName + VertStr + std::to_string(Meshes.at(CurrMesh).VertexCount) + TriCountStr + std::to_string(Meshes.at(CurrMesh).TriangleCount) + NormalCountStr + std::to_string(Meshes.at(CurrMesh).NormalCount) + MaterialNameStr + Meshes.at(CurrMesh).Mat.MaterialName + MatPtrStr, RGB(255, 0, 0), TRANSPARENT);
 //		Gdi.DrawStringA(20, 220, ResourceInfoStr + std::to_string(Material::LoadedMaterials.size()), RGB(255, 0, 0), TRANSPARENT);
-//		Gdi.DrawStringA(20, 240, CamPosXstr + std::to_string(Cam.Pos.x) + CamPosYstr + std::to_string(Cam.Pos.y) + CamPosZstr + std::to_string(Cam.Pos.z) + CamPosEndstr, RGB(255, 0, 0), TRANSPARENT);
-//		Gdi.DrawStringA(20, 260, CamRotXstr + std::to_string(Cam.ViewAngles.x) + CamRotYstr + std::to_string(Cam.ViewAngles.y) + CamRotZstr + std::to_string(Cam.ViewAngles.z) + CamRotEndstr, RGB(255, 0, 0), TRANSPARENT);
+//		Gdi.DrawStringA(20, 240, CamPosXstr + std::to_string(Cam->Pos.x) + CamPosYstr + std::to_string(Cam->Pos.y) + CamPosZstr + std::to_string(Cam->Pos.z) + CamPosEndstr, RGB(255, 0, 0), TRANSPARENT);
+//		Gdi.DrawStringA(20, 260, CamRotXstr + std::to_string(Cam->ViewAngles.x) + CamRotYstr + std::to_string(Cam->ViewAngles.y) + CamRotZstr + std::to_string(Cam->ViewAngles.z) + CamRotEndstr, RGB(255, 0, 0), TRANSPARENT);
 //	}
 //
 //	return NULL;
@@ -232,18 +229,18 @@ class ExampleScene : public TerraPGE::Scene
 	std::thread SettingsTh;
 
 	public:
-	Camera Cam;
+	Camera* Cam;
 	Texture* Txt = nullptr;
 
 	ExampleScene()
 	{
-
+		
 	}
 
 
 	TerraPGE::Renderable* SpawnCubeAt(Vec3 Position)
 	{
-		return DEBUG_NEW TerraPGE::Renderable(CubeMesh2, &Cam, Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), Position, EngineShaders::Shader_Frag_Phong_Shadows);
+		return DEBUG_NEW TerraPGE::Renderable(CubeMesh2, Cam, Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), Position, TerraPGE::EngineShaders::Shader_Frag_Phong_Shadows);
 	}
 
 
@@ -281,15 +278,15 @@ class ExampleScene : public TerraPGE::Scene
 		Dl = DirectionalLight({0, 30, -60}, dir, Vec3(253, 251, 211), 0.15f, 0.3f, 0.2f);
 		Dl.CastsShadows = true;
 
-		Cam = Camera(Vec3(0, 3, 0), (float)((float)TerraPGE::Core::sy / (float)TerraPGE::Core::sx), TerraPGE::Core::FOV, TerraPGE::Core::FNEAR, TerraPGE::Core::FFAR);
-		this->MainCamera = &Cam;
+		Cam = DEBUG_NEW Camera(Vec3(0, 3, 0), (float)((float)TerraPGE::Core::sy / (float)TerraPGE::Core::sx), TerraPGE::Core::FOV, TerraPGE::Core::FNEAR, TerraPGE::Core::FFAR);
+		this->MainCamera = Cam;
 
 		this->LoadingMode++;
 		TerraPGE::UpdateLoadingScreen();
-		CubeRender = DEBUG_NEW TerraPGE::Renderable(CubeMsh, &Cam, Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), Vec3(0.0f, 5.f, -4.0f), EngineShaders::Shader_Texture_Only);
+		CubeRender = DEBUG_NEW TerraPGE::Renderable(CubeMsh, Cam, Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), Vec3(0.0f, 5.f, -4.0f), TerraPGE::EngineShaders::Shader_Texture_Only);
 		CubeRender->collider.PhysicsEnabled = false;
-		Ak47Render = DEBUG_NEW TerraPGE::Renderable(CubeMesh2, &Cam, Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), Vec3(0.0f, 6.f, 4.0f), EngineShaders::Shader_Frag_Phong_Shadows);
-		PlaneRender = DEBUG_NEW TerraPGE::Renderable(Plane, &Cam, Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), Vec3(0.0f, 0.0f, 0.0f), EngineShaders::Shader_Frag_Phong_Shadows);
+		Ak47Render = DEBUG_NEW TerraPGE::Renderable(CubeMesh2, Cam, Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), Vec3(0.0f, 6.f, 4.0f), TerraPGE::EngineShaders::Shader_Frag_Phong_Shadows);
+		PlaneRender = DEBUG_NEW TerraPGE::Renderable(Plane, Cam, Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), Vec3(0.0f, 0.0f, 0.0f), TerraPGE::EngineShaders::Shader_Frag_Phong_Shadows);
 		CubeRender->mesh->MeshName = "SmileCube";
 
 		AABBColliderParams params = Collider::CalculateAABB(Ak47Render->mesh->Triangles);
@@ -301,7 +298,7 @@ class ExampleScene : public TerraPGE::Scene
 		PlaneRender->collider.body.KineticFriction = ICE_KINETIC_FRICTION;
 		PlaneRender->collider.PhysicsEnabled = false;
 
-		this->SlRender = DEBUG_NEW TerraPGE::Renderable(&(sl.LightMesh), &(this->Cam), Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), sl.LightPos, EngineShaders::Shader_Frag_Phong, ShaderTypes::SHADER_FRAGMENT);
+		this->SlRender = DEBUG_NEW TerraPGE::Renderable(&(sl.LightMesh), (this->Cam), Vec3(1.0f, 1.0f, 1.0f), Vec3(0.0f, 0.0f, 0.0f), sl.Transform.GetLocalPosition(), TerraPGE::EngineShaders::Shader_Frag_Phong, ShaderTypes::SHADER_FRAGMENT);
 		SettingsTh = std::thread(ExampleScene::Settings, this);
 
 		this->LoadingMode++;
@@ -319,9 +316,8 @@ class ExampleScene : public TerraPGE::Scene
 		}
 
 		LockCamera = false;
-		Cam.Transform.SetParent(&Ak47Render->Transform);
-		Cam.Transform.SetLocalPosition(Vec3(0, 2, 0));
-		Cam.PointAt(Vec3(0.0f, 6.f, 4.0f));
+		Cam->Transform.SetLocalPosition(Vec3(0, 3, 0));
+		this->Cam->Transform.SetLocalEulerAngles(Vec3(-89, 0, 0));
 	}
 
 
@@ -354,43 +350,48 @@ class ExampleScene : public TerraPGE::Scene
 	{
 		if (Wnd.Input.IsKeyDown(VK_SPACE))
 		{
-			Cam.Transform.SetLocalPosition(Cam.Transform.GetLocalPosition() + (Cam.GetNewVelocity(Vec3(0, 1, 0) * Cam.Transform.GetLocalRotationMatrix())) * ElapsedTime);
+			Cam->Transform.SetLocalPosition(Cam->Transform.GetLocalPosition() + (Cam->GetNewVelocity(Vec3(0, 1, 0) * Cam->Transform.GetLocalRotationMatrix())) * ElapsedTime);
 		}
 
 		if (Wnd.Input.IsKeyDown(VK_LSHIFT))
 		{
-			Cam.Transform.SetLocalPosition(Cam.Transform.GetLocalPosition() + (Cam.GetNewVelocity(Vec3(0, -1, 0) * Cam.Transform.GetLocalRotationMatrix())) * ElapsedTime);
+			Cam->Transform.SetLocalPosition(Cam->Transform.GetLocalPosition() + (Cam->GetNewVelocity(Vec3(0, -1, 0) * Cam->Transform.GetLocalRotationMatrix())) * ElapsedTime);
 		}
 
 		if (Wnd.Input.IsKeyDown('A'))
 		{
-			Cam.Transform.SetLocalPosition(Cam.Transform.GetLocalPosition() + (Cam.GetNewVelocity(Vec3(-1, 0, 0) * Cam.Transform.GetLocalRotationMatrix())) * ElapsedTime);
+			Cam->Transform.SetLocalPosition(Cam->Transform.GetLocalPosition() + (Cam->GetNewVelocity(Vec3(-1, 0, 0) * Cam->Transform.GetLocalRotationMatrix())) * ElapsedTime);
 		}
 
 		if (Wnd.Input.IsKeyDown('D'))
 		{
-			Cam.Transform.SetLocalPosition(Cam.Transform.GetLocalPosition() + (Cam.GetNewVelocity(Vec3(1, 0, 0) * Cam.Transform.GetLocalRotationMatrix())) * ElapsedTime);
+			Cam->Transform.SetLocalPosition(Cam->Transform.GetLocalPosition() + (Cam->GetNewVelocity(Vec3(1, 0, 0) * Cam->Transform.GetLocalRotationMatrix())) * ElapsedTime);
 		}
 
 		if (Wnd.Input.IsKeyDown('W'))
 		{
-			Cam.Transform.SetLocalPosition(Cam.Transform.GetLocalPosition() + (Cam.GetNewVelocity(Vec3(0, 0, 1) * Cam.Transform.GetLocalRotationMatrix())) * ElapsedTime);
+			Cam->Transform.SetLocalPosition(Cam->Transform.GetLocalPosition() + (Cam->GetNewVelocity(Vec3(0, 0, 1) * Cam->Transform.GetLocalRotationMatrix())) * ElapsedTime);
 		}
 
 		if (Wnd.Input.IsKeyDown('S'))
 		{
-			Cam.Transform.SetLocalPosition(Cam.Transform.GetLocalPosition() + (Cam.GetNewVelocity(Vec3(0, 0, -1) * Cam.Transform.GetLocalRotationMatrix())) * ElapsedTime);
+			Cam->Transform.SetLocalPosition(Cam->Transform.GetLocalPosition() + (Cam->GetNewVelocity(Vec3(0, 0, -1) * Cam->Transform.GetLocalRotationMatrix())) * ElapsedTime);
+		}
+
+		if (Wnd.Input.IsKeyDown('C'))
+		{
+			CubeRender->Transform.WalkTransformChain();
+			Cam->Transform.PointAt(CubeRender->Transform.GetWorldPosition(), Cam->CamUp);
 		}
 
 		if (!LockCamera)
 		{
-			Vec3 Euler = Cam.Transform.GetLocalEulerAngles();
+			Vec3 Euler = Cam->Transform.GetLocalEulerAngles();
 
-			Euler += Vec3(-1 * (float)Wnd.Input.Current.DeltaY * TerraPGE::Sensitivity, 0, 0);
-			Euler.x = std::clamp<float>(Euler.x, -89.0f, 89.0f);
+			Euler += Vec3((float)Wnd.Input.Current.DeltaY * TerraPGE::Sensitivity, 0, 0);
 			Euler -= Vec3(0, (float)Wnd.Input.Current.DeltaX * TerraPGE::Sensitivity, 0);
 
-			Cam.Transform.SetLocalEulerAngles(Euler);
+			Cam->Transform.SetLocalEulerAngles(Euler.AngleNormalized());
 		}
 
 	}
@@ -401,9 +402,10 @@ class ExampleScene : public TerraPGE::Scene
 		TerraPGE::Renderable* hovered = nullptr;
 		for (TerraPGE::Renderable* obj : *ToRender)
 		{
+			Cam->Transform.WalkTransformChain();
 			Matrix inv = obj->Transform.World.QuickInversed();
-			Vec3 localOrigin = Cam.Transform.GetLocalPosition() * inv;
-			Vec3 localDir = ((Cam.GetLookDirection()) * inv.ExtractRotationScale()).Normalized();
+			Vec3 localOrigin = Cam->Transform.GetWorldPosition() * inv;
+			Vec3 localDir = ((Cam->GetLookDirection()) * inv.ExtractRotationScale()).Normalized();
 
 
 			if (!obj) continue;
@@ -423,12 +425,12 @@ class ExampleScene : public TerraPGE::Scene
 
 	void RunTick(GdiPP* Gdi, WndCreator& Wnd, const float& ElapsedTime, std::vector<TerraPGE::Renderable*>* ToRender, std::vector<LightObject*>* Lights) override
 	{
-		static float PrevFov = Cam.Fov;
+		static float PrevFov = Cam->Fov;
 
-		if (PrevFov != Cam.Fov)
+		if (PrevFov != Cam->Fov)
 		{
-			Cam.CalcProjectionMat();
-			PrevFov = Cam.Fov;
+			Cam->CalcProjectionMat();
+			PrevFov = Cam->Fov;
 		}
 
 		if (Wnd.Input.IsKeyPressed(VK_INSERT))
@@ -446,7 +448,7 @@ class ExampleScene : public TerraPGE::Scene
 				IsFullScreen = !IsFullScreen;
 			}
 			TerraPGE::Core::UpdateScreenInfo(Gdi);
-			Cam.AspectRatio = (float)((float)TerraPGE::Core::sy / (float)TerraPGE::Core::sx);
+			Cam->AspectRatio = (float)((float)TerraPGE::Core::sy / (float)TerraPGE::Core::sx);
 		}
 
 		if (Wnd.Input.IsKeyPressed(VK_ESCAPE))
@@ -460,12 +462,12 @@ class ExampleScene : public TerraPGE::Scene
 
 		if (Wnd.Input.IsKeyPressed(VK_HOME))
 		{
-			LockCamera = !LockCamera;
-			Matrix Proj;
-			Proj.MakeOrthoMatrix(-40.0f, 40.0f, -40.0f, 40.0f, TerraPGE::Core::FNEAR, TerraPGE::Core::FFAR);
-			Cam.ProjectionMatrix = Proj;
-			Cam.ViewMatrix = Matrix::CalcViewMatrix(((Vec3(0.0f, 0.0f, 0.0f) - LightSrcPos).Normalized()) * 100.0f, Vec3(0.0f, 0.0f, 0.0f), Vec3(0, 1, 0));
-			Cam.Transform.SetLocalPosition(-((Vec3(0.0f, 0.0f, 0.0f) - LightSrcPos).Normalized()) * 100.0f);
+			//LockCamera = !LockCamera;
+			//Matrix Proj;
+			//Proj.MakeOrthoMatrix(-40.0f, 40.0f, -40.0f, 40.0f, TerraPGE::Core::FNEAR, TerraPGE::Core::FFAR);
+			//Cam->ProjectionMatrix = Proj;
+			//Cam->ViewMatrix = Matrix::CalcViewMatrix(((Vec3(0.0f, 0.0f, 0.0f) - LightSrcPos).Normalized()) * 100.0f, Vec3(0.0f, 0.0f, 0.0f), Vec3(0, 1, 0));
+			//Cam->Transform.SetLocalPosition(-((Vec3(0.0f, 0.0f, 0.0f) - LightSrcPos).Normalized()) * 100.0f);
 		}
 
 		if (Wnd.Input.IsKeyDown('P'))
@@ -477,7 +479,7 @@ class ExampleScene : public TerraPGE::Scene
 		{
 			if (HoveredRend)
 			{
-				HoveredRend->collider.body.Velocity = Cam.GetLookDirection().Normalized() * 8.0f;
+				HoveredRend->collider.body.Velocity = Cam->GetLookDirection().Normalized() * 8.0f;
 			}
 		}
 
@@ -567,8 +569,8 @@ class ExampleScene : public TerraPGE::Scene
 		if (TerraPGE::Core::FpsEngineCounter && ShowStrs)
 		{
 			static std::string FovStr = "(F2)  Fov: ";
-			static std::string YawRotStr = "(F3)  Yaw Rotation Speed: ";
-			static std::string PitchRotStr = "(F4)  Pitch Rotation Speed: ";
+			static std::string YawRotStr = "(F3)  Pitch Rotation Speed: ";
+			static std::string PitchRotStr = "(F4)  Yaw Rotation Speed: ";
 			static std::string RollRotStr = "(F5)  Roll Rotation Speed: ";
 			static std::string CullingStr = "(F6)  Culling: ";
 			static std::string LightingStr = "(F7)  Lighting: ";
@@ -583,12 +585,17 @@ class ExampleScene : public TerraPGE::Scene
 			static std::string CamPosYstr = ", Y: ";
 			static std::string CamPosZstr = ", Z: ";
 			static std::string CamPosEndstr = ")";
-			static std::string CamRotXstr = "Camera Rot: ( Yaw: ";
-			static std::string CamRotYstr = ", Pitch: ";
+			static std::string CamRotXstr = "Camera Rot: ( Pitch: ";
+			static std::string CamRotYstr = ", Yaw: ";
 			static std::string CamRotZstr = ", Roll: ";
 			static std::string CamRotEndstr = ")";
 			static std::string ResourceInfoStr = "LoadedMaterials: ";
 			static std::string ResourceInfoEndStr = "";
+			static std::string DepthStr = "Depth: ";
+			static std::string ClipStr = "Clip: (z, w): ";
+			static std::string NdcStr = "Ndc: ";
+
+
 			std::stringstream ss;
 			//ss << " : 0x"; << std::hex << &Meshes.at(CurrMesh).Mat << std::dec;
 			std::string MatPtrStr = ss.str();
@@ -599,8 +606,12 @@ class ExampleScene : public TerraPGE::Scene
 			Gdi->DrawStringA(20, 100, FilledStr + std::to_string(TerraPGE::Core::WireFrame), RGB(255, 0, 0), TRANSPARENT);
 			Gdi->DrawStringA(20, 120, TriLinesStr + std::to_string(TerraPGE::Core::ShowTriLines), RGB(255, 0, 0), TRANSPARENT);
 			Gdi->DrawStringA(20, 140, ResourceInfoStr + std::to_string(Material::LoadedMaterials.size()), RGB(255, 0, 0), TRANSPARENT);
-			Gdi->DrawStringA(20, 160, CamPosXstr + std::to_string(Cam.Transform.GetLocalPosition().x) + CamPosYstr + std::to_string(Cam.Transform.GetLocalPosition().y) + CamPosZstr + std::to_string(Cam.Transform.GetLocalPosition().z) + CamPosEndstr, RGB(255, 0, 0), TRANSPARENT);
-			Gdi->DrawStringA(20, 180, CamRotXstr + std::to_string(Cam.Transform.GetLocalEulerAngles().x) + CamRotYstr + std::to_string(Cam.Transform.GetLocalEulerAngles().y) + CamRotZstr + std::to_string(Cam.Transform.GetLocalEulerAngles().z) + CamRotEndstr, RGB(255, 0, 0), TRANSPARENT);
+			Gdi->DrawStringA(20, 160, CamPosXstr + std::to_string(Cam->Transform.GetLocalPosition().x) + CamPosYstr + std::to_string(Cam->Transform.GetLocalPosition().y) + CamPosZstr + std::to_string(Cam->Transform.GetLocalPosition().z) + CamPosEndstr, RGB(255, 0, 0), TRANSPARENT);
+			Gdi->DrawStringA(20, 180, CamRotXstr + std::to_string(Cam->Transform.GetLocalEulerAngles().x) + CamRotYstr + std::to_string(Cam->Transform.GetLocalEulerAngles().y) + CamRotZstr + std::to_string(Cam->Transform.GetLocalEulerAngles().z) + CamRotEndstr, RGB(255, 0, 0), TRANSPARENT);
+			Gdi->DrawStringA(20, 220, DepthStr + std::to_string(TerraPGE::Renderer::TestDepth), RGB(255, 0, 0), TRANSPARENT);
+			Gdi->DrawStringA(20, 240, ClipStr + std::to_string(TerraPGE::Renderer::TestClipZ) + ", " + std::to_string(TerraPGE::Renderer::TestClipW), RGB(255, 0, 0), TRANSPARENT);
+			Gdi->DrawStringA(20, 260, NdcStr + std::to_string(TerraPGE::Renderer::TestNdcZ), RGB(255, 0, 0), TRANSPARENT);
+
 		}
 
 		DrawCrosshair(Gdi);
@@ -629,11 +640,11 @@ class ExampleScene : public TerraPGE::Scene
 			}
 			if (GetAsyncKeyState(VK_F2))
 			{
-				Scene->Cam.Fov += 30;
+				Scene->Cam->Fov += 30;
 				TerraPGE::Core::FOV += 30;
 				if (TerraPGE::Core::FOV > 120)
 				{
-					Scene->Cam.Fov = 60;
+					Scene->Cam->Fov = 60;
 					TerraPGE::Core::FOV = 60;
 				}
 			}
